client/repos: refactor repos page and repodetails page to use codegen queries
